requirements.txt
streamlit>=1.20
snscrape>=0.5.0
pandas>=2.0
python-dotenv>=0.21
plotly>=5.15.0
plotly.express>=0.4.1
nltk>=3.8.1